remove depenency to ember-data and reduce packages that were not needed
- remove ember-data dependency by not providing an example router and controller. It is better to be explicit about this and show the code needed to the user in the tutorial. This makes it more useful to build further upon.
Update dummy app to not use these.

- only depend on the specific lodash package that is used.

- reference the serializer in the readme

- reference the dummy app in the readme

- remove dependency ember-math-helpers. This was unused
