+++ package.json
@@ -33,10 +33,8 @@
     "ember-auto-import": "^2.0.0",
     "ember-cli-babel": "^7.26.10",
     "ember-cli-htmlbars": "^5.7.2",
-    "ember-data": ">=3.28.0",
-    "ember-math-helpers": "^2.18.0",
     "ember-truth-helpers": "^3.0.0",
-    "lodash": "^4.17.21"
+    "lodash.upperfirst": "^4.3.1"
   },
   "devDependencies": {
     "@ember/optional-features": "^2.0.0",
@@ -46,7 +44,6 @@
     "@glimmer/tracking": "^1.0.4",
     "babel-eslint": "^10.1.0",
     "broccoli-asset-rev": "^3.0.0",
-    "ember-cached-decorator-polyfill": "^1.0.2",
     "ember-cli": "~3.28.5",
     "ember-cli-dependency-checker": "^3.2.0",
     "ember-cli-inject-live-reload": "^2.1.0",
@@ -55,6 +52,7 @@
     "ember-cli-sri": "^2.1.1",
     "ember-cli-terser": "^4.0.2",
     "ember-cli-update": "^0.33.2",
+    "ember-data": ">=3.28.0",
     "ember-disable-prototype-extensions": "^1.1.3",
     "ember-export-application-global": "^2.0.1",
     "ember-load-initializers": "^2.1.2",
