+++ tests/dummy/app/controllers/adv-table-ed.js
@@ -1,3 +1,12 @@
-import DataTableController from 'ember-data-table/controller';
+import { tracked } from '@glimmer/tracking';
+import Controller from '@ember/controller';
 
-export default class AdvTableEdController extends DataTableController {}
+export default class AdvTableEdController extends Controller {
+  queryParams = ['size', 'page', 'filter', 'sort'];
+
+  @tracked size = 10;
+  @tracked page = 0;
+  @tracked filter = '';
+  @tracked sort = '';
+  @tracked isLoadingModel = false;
+}
