+++ tests/dummy/app/controllers/adv-table-pojo.js
@@ -1,3 +1,12 @@
-import DataTableController from 'ember-data-table/controller';
+import { tracked } from '@glimmer/tracking';
+import Controller from '@ember/controller';
 
-export default class AdvTablePojoController extends DataTableController {}
+export default class AdvTablePojoController extends Controller {
+  queryParams = ['size', 'page', 'filter', 'sort'];
+
+  @tracked size = 10;
+  @tracked page = 0;
+  @tracked filter = '';
+  @tracked sort = '';
+  @tracked isLoadingModel = false;
+}
