+++ tests/dummy/app/controllers/simple-table-ed.js
@@ -1,3 +1,11 @@
-import DataTableController from 'ember-data-table/controller';
+import Controller from '@ember/controller';
+import { tracked } from '@glimmer/tracking';
+export default class SimpleTableEdController extends Controller {
+  queryParams = ['size', 'page', 'filter', 'sort'];
 
-export default class SimpleTableEdController extends DataTableController {}
+  @tracked size = 10;
+  @tracked page = 0;
+  @tracked filter = '';
+  @tracked sort = '';
+  @tracked isLoadingModel = false;
+}
