+++ tests/dummy/app/controllers/simple-table-pojo.js
@@ -1,3 +1,12 @@
-import DataTableController from 'ember-data-table/controller';
+import Controller from '@ember/controller';
+import { tracked } from '@glimmer/tracking';
 
-export default class SimpleTablePojoController extends DataTableController {}
+export default class SimpleTablePojoController extends Controller {
+  queryParams = ['size', 'page', 'filter', 'sort'];
+
+  @tracked size = 10;
+  @tracked page = 0;
+  @tracked filter = '';
+  @tracked sort = '';
+  @tracked isLoadingModel = false;
+}
