+++ tests/dummy/app/templates/adv-table-pojo.hbs
@@ -1,5 +1,4 @@
 {{page-title 'Advanced table + objects'}}
-test
 <div class='layout-row flex layout-align-center'>
   <div class='container'>
     <h1 id='title'>
